pull apart logic in typeahead to organize things a bit better since this has gained a lot of complexity from the initial structure! styled components live in components.js to fit with our existing pattern, new useTypeahead file to memoized state in the same structure ast stories and templates. useEffects that do not interact with the actual component live there too. Now only refs and callbacks are necessary in the actual typeaheadInput file
