+++ assets/src/dashboard/components/index.js
@@ -75,8 +75,7 @@ export {
 } from './table';
 export { Toaster, ToastProvider, useToastContext } from './toaster';
 export { default as ToggleButtonGroup } from './toggleButtonGroup';
-export { default as TypeaheadInput } from './typeaheadInput';
-export { default as TypeaheadOptions } from './typeaheadOptions';
+export { TypeaheadInput, TypeaheadOptions } from './typeahead';
 export {
   DefaultParagraph1,
   Heading1,
